rename main module to window module
