+++ test/manufacturing.test.cpp
@@ -1,4 +1,4 @@
-#include "modules/main/main_module.h"
+#include "modules/window/window_module.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
 #include "modules/site/site.h"
@@ -8,7 +8,7 @@
 SCENARIO("systemBuildingUpdateManufacuringProgress", "[system]") {
   flecs::world world;
   world.import <SimulationModule>();
-  world.import <MainModule>();
+  world.import <WindowModule>();
   world.import <RocketModule>();
   world.import <SiteModule>();
