+++ test/rocket/actions/rocket_complete_move_action.test.cpp
@@ -1,4 +1,4 @@
-#include "modules/main/main_module.h"
+#include "modules/window/window_module.h"
 #include "modules/rocket/actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
@@ -8,7 +8,7 @@
 SCENARIO("RocketCompleteMoveAction", "[action]") {
   flecs::world world;
   world.import <SimulationModule>();
-  world.import <MainModule>();
+  world.import <WindowModule>();
   world.import <RocketModule>();
 
   GIVEN("An invalid rocket entity") {
@@ -107,7 +107,7 @@ SCENARIO("RocketCompleteMoveAction", "[action]") {
 SCENARIO("RocketCompleteMoveAction Block", "[action]") {
   flecs::world world;
   world.import <SimulationModule>();
-  world.import <MainModule>();
+  world.import <WindowModule>();
   world.import <RocketModule>();
 
   GIVEN("A rocket that can complete move") {
