+++ test/rocket/systems.test.cpp
@@ -1,5 +1,5 @@
 #include "modules/engine/render.h"
-#include "modules/main/main_module.h"
+#include "modules/window/window_module.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
 #include "modules/site/site.h"
@@ -199,7 +199,7 @@ SCENARIO("systemLaunchRocket", "[rocket][system]") {
 SCENARIO("systemRocketCompleteAction", "[system]") {
   flecs::world world;
   world.import <SimulationModule>();
-  world.import <MainModule>();
+  world.import <WindowModule>();
   world.import <RocketModule>();
 
   GIVEN("A rocket under construction with effort still remaining") {
